requirements.txt
asgiref==3.4.1
attrs==21.2.0
coverage==5.5
dj-database-url==0.5.0
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
factory-boy==3.2.0
Faker==8.12.1
gunicorn==20.1.0
inflection==0.5.1
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
whitenoise==5.3.0
